fix: ready for deployment bugs fix
